chore: made multiprocess test a little more reliable

+++ test/tap/multiprocess-shutdown-test.js
@@ -119,18 +119,18 @@ test('multiprocess appender crash (worker)', (t) => {
     categories: { default: { appenders: ['multi'], level: 'debug' } }
   });
 
-  setTimeout(() => {
-    const worker = childProcess.fork(
-      require.resolve('./multiprocess-worker'),
-      ['start-multiprocess-worker', loggerPort]
-    );
+  const worker = childProcess.fork(
+    require.resolve('./multiprocess-worker'),
+    ['start-multiprocess-worker', loggerPort]
+  );
 
-    setTimeout(() => {
+  worker.on('message', (m) => {
+    if (m === 'worker is done') {
       worker.kill();
       setTimeout(() => {
         t.equal(messages[0], 'Logging from worker');
         log4jsWithFakeConsole.shutdown(() => t.end());
-      }, 250);
-    }, 250);
-  }, 250);
+      }, 500);
+    }
+  });
 });
